# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "IncidentDesk",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/v1/services": {
      "get": {
        "summary": "Accessible services",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        }
      }
    },
...
    "/internal/v1/source-change": {
      "post": {
        "summary": "Trusted source notification; X-Source-Key identity",
        "security": [
          {
            "SourceKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Durably accepted"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Scope or role forbidden"
          },
          "409": {
            "description": "Version, lease or idempotency conflict"
          },
          "429": {
            "description": "Quota exceeded"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "service": {
                    "type": "string"
                  },
                  "delete": {
                    "type": "boolean"
                  }
                },
                "required": [
                  "service",
                  "delete"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "name": "X-Protocol-Version",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "1"
              ]
            }
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "OIDC JWT"
      },
      "SourceKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Source-Key"
      },
      "WorkerKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Worker-Key"
      }
    }
  }
}